[refinement estimation] refactoring: add PerSkeletonRefinementEstimator parent class
