+++ predicators/settings.py
@@ -277,6 +277,7 @@ class GlobalSettings:
     doors_draw_debug = False
 
     # narrow_passage env parameters
+    narrow_passage_open_door_refine_penalty = 0.2
     narrow_passage_door_width_padding_lb = 1e-4
     narrow_passage_door_width_padding_ub = 0.015
     narrow_passage_passage_width_padding_lb = 5e-4
