+++ tests/approaches/test_refinement_estimation_approach.py
@@ -26,6 +26,7 @@ def test_refinement_estimation_approach():
         "narrow_passage_door_width_padding_ub": 0.05,
         "narrow_passage_passage_width_padding_lb": 0.05,
         "narrow_passage_passage_width_padding_ub": 0.05,
+        "narrow_passage_open_door_refine_penalty": 0,
     }
     # Default to 2 train and test tasks, but allow them to be specified in
     # the extra args too.
